refactor: break up into smaller files
