+++ packages/pure-parse/src/parsing/optionalSymbol.ts
@@ -1,0 +1,4 @@
+/**
+ * Used to represent optional validators at runtime and compile-time in two different ways
+ */
+export const optionalSymbol = Symbol('optional parser')
