+++ packages/pure-parse/src/parsing/union.test.ts
@@ -1,0 +1,8 @@
+import { describe } from 'vitest'
+
+describe('unions', () => {
+  describe.todo('union')
+  describe.todo('nullable')
+  describe.todo('undefinable')
+  describe.todo('optionalNullable')
+})
