+++ packages/pure-parse/src/shared/primitive.ts
@@ -1,0 +1,11 @@
+/**
+ * A JavaScript primitive
+ */
+export type Primitive =
+  | null
+  | undefined
+  | boolean
+  | number
+  | string
+  | bigint
+  | symbol
